+++ Reborn-Zune-Common/Reborn-Zune-Common.csproj
@@ -120,8 +120,9 @@
     <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Class1.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Services\IService.cs" />
+    <Compile Include="Services\ServiceLocator.cs" />
     <EmbeddedResource Include="Properties\Reborn_Zune_Common.rd.xml" />
   </ItemGroup>
   <ItemGroup>
